breakout devices into own files
